Cast target to plain int in trainers — survives BQ's nullable Int64
When training data comes from BigQuery via to_dataframe(), an INT64
column arrives as pandas nullable Int64 (capital I). sklearn promotes
that to float64 internally, classes_ ends up as float, predict()
returns floats. Same with LightGBM. Reading the same data from CSV
sees plain int64 and predictions stay as int.

`y = df["target"].astype(int)` collapses both nullable and plain
integer dtypes to the same numpy int64, keeping classes_ integer-typed
and predictions int. Verified end-to-end via the BQ exploration
notebook: predictions for sonar now come back as [0, 0, 0, 0, 0]
instead of [0.0, 0.0, 0.0, 0.0, 0.0].

Applied to both src/train.py (sklearn) and src/train_lightgbm.py.
src/train_torch.py already does astype("int64") explicitly.

+++ src/train.py
@@ -77,7 +77,9 @@ def main():
     # drop bookkeeping columns Feast adds (entity + timestamp) if present
     feature_cols = [c for c in df.columns if c not in {"target", "signal_id", "event_timestamp"}]
     X = df[feature_cols]
-    y = df["target"]
+    # Cast to plain int so sklearn doesn't promote classes_ to float when
+    # `target` arrives as pandas nullable Int64 (e.g. from BigQuery).
+    y = df["target"].astype(int)
     X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
 
     run_params = {"n-estimators": args.n_estimators, "max-depth": args.max_depth,
